Added `speak_messages` Python control for OS mode

+++ interpreter/terminal_interface/terminal_interface.py
@@ -213,38 +213,45 @@ def terminal_interface(interpreter, message):
 
                     if "end" in chunk and interpreter.os:
                         last_message = interpreter.messages[-1]["content"]
-                        if (
-                            platform.system() == "Darwin"
-                            and last_message not in force_task_completion_responses
-                        ):
-                            # Remove markdown lists and the line above markdown lists
-                            lines = last_message.split("\n")
-                            i = 0
-                            while i < len(lines):
-                                # Match markdown lists starting with hyphen, asterisk or number
-                                if re.match(r"^\s*([-*]|\d+\.)\s", lines[i]):
-                                    del lines[i]
-                                    if i > 0:
-                                        del lines[i - 1]
-                                        i -= 1
-                                else:
-                                    i += 1
-                            message = "\n".join(lines)
-                            # Replace newlines with spaces, escape double quotes and backslashes
-                            sanitized_message = (
-                                message.replace("\\", "\\\\")
-                                .replace("\n", " ")
-                                .replace('"', '\\"')
-                            )
-                            if voice_subprocess:
-                                voice_subprocess.terminate()
-                            voice_subprocess = subprocess.Popen(
-                                [
-                                    "osascript",
-                                    "-e",
-                                    f'say "{sanitized_message}" using "Fred"',
-                                ]
-                            )
+                        # Useful for OS mode— how do we display messages without the terminal being displayed?
+                        # IN THE FUTURE I think this should be notifications, not voice.
+                        if interpreter.speak_messages:
+                            if (
+                                platform.system() == "Darwin"
+                                and last_message not in force_task_completion_responses
+                            ):
+                                # Remove markdown lists and the line above markdown lists
+                                lines = last_message.split("\n")
+                                i = 0
+                                while i < len(lines):
+                                    # Match markdown lists starting with hyphen, asterisk or number
+                                    if re.match(r"^\s*([-*]|\d+\.)\s", lines[i]):
+                                        del lines[i]
+                                        if i > 0:
+                                            del lines[i - 1]
+                                            i -= 1
+                                    else:
+                                        i += 1
+                                message = "\n".join(lines)
+                                # Replace newlines with spaces, escape double quotes and backslashes
+                                sanitized_message = (
+                                    message.replace("\\", "\\\\")
+                                    .replace("\n", " ")
+                                    .replace('"', '\\"')
+                                )
+                                if voice_subprocess:
+                                    voice_subprocess.terminate()
+                                voice_subprocess = subprocess.Popen(
+                                    [
+                                        "osascript",
+                                        "-e",
+                                        f'say "{sanitized_message}" using "Fred"',
+                                    ]
+                                )
+                            else:
+                                pass
+                                # User isn't on a Mac, so we can't do this. You should tell them something about that when they first set this up.
+                                # Or use a universal TTS library.
 
                 # Assistant code blocks
                 elif chunk["role"] == "assistant" and chunk["type"] == "code":
